seggregate create_time_series_factory
