+++ source/databricks/settlement_report/settlement_report_job/domain/report_generator.py
@@ -9,7 +9,7 @@ from settlement_report_job.domain.DataProductValues.metering_point_resolution im
 from settlement_report_job.domain.repository import WholesaleRepository
 from settlement_report_job.domain.report_data_type import ReportDataType
 from settlement_report_job.domain.settlement_report_args import SettlementReportArgs
-from settlement_report_job.domain.time_series_factory import (
+from settlement_report_job.domain.time_series.time_series_factory import (
     create_time_series_for_wholesale,
 )
 from settlement_report_job.domain.task_type import TaskType
